dishesz/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
boto3==1.25.3
botocore==1.28.3
channels==3.0.4
Django==4.1.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.1
iniconfig==1.1.1
jmespath==1.0.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
mysqlclient
daphne
channels_redis==3.3.1
numpy
pandas
pytest-asyncio 
anyio
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted
requests
progress


